refactor(core): Add factory functions to create instances of IRunExecutionData
We have instantiated objects of `IRunExecutionData` all over the codebase. To
have central control of that type, this PR encapsulated the creation of instances
of this interface. This makes it easier to track where they are created and for what
use cases. Also makes it easier to make changes to the interface.

The main motivation for this is that I want to add a new property to the interface
in a coming up PR.